neurooracle/requirements.txt
sentence-transformers
networkx
umap-learn
hdbscan
bertopic
matplotlib
plotly
duckdb
faiss-cpu
pandas
scikit-learn
